# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arpit Thrift ecommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  images: string[];
  category: string;
  sizes?: string[];
  colors?: string[];
  inStock: boolean;
  /** @nullable */
  stockCount?: number | null;
  isNew: boolean;
  isBestSeller: boolean;
  isFeatured: boolean;
  tags?: string[];
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  images: string[];
  category: string;
  sizes?: string[];
  colors?: string[];
  inStock?: boolean;
  /** @nullable */
  stockCount?: number | null;
  isNew?: boolean;
  isBestSeller?: boolean;
  isFeatured?: boolean;
  tags?: string[];
}

export interface ProductUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  images?: string[];
  category?: string;
  sizes?: string[];
  colors?: string[];
  inStock?: boolean;
  /** @nullable */
  stockCount?: number | null;
  isNew?: boolean;
  isBestSeller?: boolean;
  isFeatured?: boolean;
  tags?: string[];
}

export interface FeaturedProducts {
  featured: Product[];
  newArrivals: Product[];
  bestSellers: Product[];
}

export interface OrderItem {
  productId: number;
  productName: string;
  productImage: string;
  price: number;
  quantity: number;
  /** @nullable */
  size: string | null;
  /** @nullable */
  color: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending_payment: 'pending_payment',
  payment_verified: 'payment_verified',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  orderId: string;
  fullName: string;
  phone: string;
  address: string;
  city: string;
  state: string;
  pincode: string;
  items: OrderItem[];
  totalAmount: number;
  status: OrderStatus;
  paymentStatus: string;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  razorpayPaymentId?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  fullName: string;
  phone: string;
  address: string;
  city: string;
  state: string;
  pincode: string;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending_payment: 'pending_payment',
  payment_verified: 'payment_verified',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface PaymentOrderInput {
  orderId: number;
}

export interface PaymentOrderResponse {
  razorpayOrderId: string;
  amount: number;
  currency: string;
  keyId: string;
  orderId: number;
  orderRef: string;
  customerName: string;
  customerPhone: string;
}

export interface PaymentVerifyInput {
  razorpay_order_id: string;
  razorpay_payment_id: string;
  razorpay_signature: string;
  orderId: number;
}

export interface PaymentVerifyResponse {
  success: boolean;
  orderId: number;
  orderRef: string;
}

export interface AdminStats {
  totalOrders: number;
  pendingPayment: number;
  paymentVerified: number;
  shipped: number;
  delivered: number;
  totalRevenue: number;
  totalProducts: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
sortBy?: ListProductsSortBy;
inStock?: boolean;
};

export type ListProductsSortBy = typeof ListProductsSortBy[keyof typeof ListProductsSortBy];


export const ListProductsSortBy = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  newest: 'newest',
  popular: 'popular',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending_payment: 'pending_payment',
  payment_verified: 'payment_verified',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;