# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arpit Thrift ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalOrders: number;
  pendingPayment: number;
  paymentVerified: number;
  shipped: number;
  delivered: number;
  totalRevenue: number;
  totalProducts: number;
}