# lib/api-zod/src/generated/types/featuredProducts.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arpit Thrift ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface FeaturedProducts {
  featured: Product[];
  newArrivals: Product[];
  bestSellers: Product[];
}