# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arpit Thrift ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderId: string;
  fullName: string;
  phone: string;
  address: string;
  city: string;
  state: string;
  pincode: string;
  items: OrderItem[];
  totalAmount: number;
  status: OrderStatus;
  paymentStatus: string;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  razorpayPaymentId?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}