# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arpit Thrift ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: number;
  productName: string;
  productImage: string;
  price: number;
  quantity: number;
  /** @nullable */
  size: string | null;
  /** @nullable */
  color: string | null;
}