# lib/api-zod/src/generated/types/orderStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arpit Thrift ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatusUpdateStatus } from './orderStatusUpdateStatus';

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}