# lib/api-zod/src/generated/types/paymentVerifyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arpit Thrift ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentVerifyInput {
  razorpay_order_id: string;
  razorpay_payment_id: string;
  razorpay_signature: string;
  orderId: number;
}